PY-105 Move alpha.filters to internal.filters
`alpha.filters` contain both types/API and an implementation. There's no
real separation right now, so let's move the current version to
internal, so that `neptune_fetcher.alpha.filters` becomes
`neptune_fetcher.internal.filters`.

`neptune_fetcher.alpha.filters` aliases the internal filters with the
option to futher develop more filters without the need to expose all the
changes to alpha.
