+++ src/neptune_fetcher/internal/composition/attributes.py
@@ -24,7 +24,7 @@ from typing import (
 
 from neptune_api.client import AuthenticatedClient
 
-import neptune_fetcher.alpha.filters as filters
+import neptune_fetcher.internal.filters as filters
 from neptune_fetcher.internal import (
     env,
     identifiers,
