+++ src/neptune_fetcher/internal/composition/fetch_metrics.py
@@ -27,7 +27,7 @@ from typing import (
 import pandas as pd
 from neptune_api.client import AuthenticatedClient
 
-from neptune_fetcher.alpha.filters import (
+from neptune_fetcher.internal.filters import (
     AttributeFilter,
     Filter,
 )
