+++ src/neptune_fetcher/internal/composition/fetch_table.py
@@ -22,7 +22,7 @@ from typing import (
 
 import pandas as pd
 
-from neptune_fetcher.alpha.filters import (
+from neptune_fetcher.internal.filters import (
     Attribute,
     AttributeFilter,
     Filter,
