+++ src/neptune_fetcher/internal/composition/list_attributes.py
@@ -21,7 +21,7 @@ from typing import (
     Optional,
 )
 
-from neptune_fetcher.alpha.filters import (
+from neptune_fetcher.internal.filters import (
     AttributeFilter,
     Filter,
 )
