+++ src/neptune_fetcher/internal/composition/list_containers.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 from typing import Optional
 
-from neptune_fetcher.alpha.filters import Filter
+from neptune_fetcher.internal.filters import Filter
 from neptune_fetcher.internal import client as _client
 from neptune_fetcher.internal import context as _context
 from neptune_fetcher.internal import identifiers
