+++ src/neptune_fetcher/internal/util.py
@@ -18,7 +18,7 @@ from typing import (
     Union,
 )
 
-from neptune_fetcher.alpha import filters as _filters
+from neptune_fetcher.internal import filters as _filters
 
 
 def resolve_runs_filter(runs: Optional[Union[str, list[str], _filters.Filter]]) -> Optional[_filters.Filter]:
