+++ tests/e2e/alpha/internal/composition/test_attributes.py
@@ -9,7 +9,7 @@ from datetime import (
 
 import pytest
 
-from neptune_fetcher.alpha.filters import AttributeFilter
+from neptune_fetcher.internal.filters import AttributeFilter
 from neptune_fetcher.internal.composition.attributes import fetch_attribute_definitions
 from neptune_fetcher.internal.identifiers import RunIdentifier
 from neptune_fetcher.internal.retrieval.attribute_definitions import AttributeDefinition
@@ -29,7 +29,7 @@ def run_with_attributes(client, project):
 
     from neptune_scale import Run
 
-    from neptune_fetcher.alpha.filters import Filter
+    from neptune_fetcher.internal.filters import Filter
     from neptune_fetcher.internal import identifiers
     from neptune_fetcher.internal.retrieval.search import fetch_experiment_sys_attrs
 
@@ -84,7 +84,7 @@ def run_with_attributes(client, project):
 
 @pytest.fixture(scope="module")
 def experiment_identifier(client, project, run_with_attributes) -> RunIdentifier:
-    from neptune_fetcher.alpha.filters import Filter
+    from neptune_fetcher.internal.filters import Filter
     from neptune_fetcher.internal.retrieval.search import fetch_experiment_sys_attrs
 
     project_identifier = project.project_identifier
