+++ tests/e2e/alpha/internal/composition/test_type_inference.py
@@ -7,7 +7,7 @@ from datetime import (
 import pytest
 
 from neptune_fetcher.alpha.exceptions import AttributeTypeInferenceError
-from neptune_fetcher.alpha.filters import (
+from neptune_fetcher.internal.filters import (
     Attribute,
     Filter,
 )
