+++ tests/e2e/alpha/internal/test_fetch_metrics.py
@@ -14,7 +14,7 @@ import pandas as pd
 import pytest
 
 from neptune_fetcher.alpha import fetch_metrics
-from neptune_fetcher.alpha.filters import (
+from neptune_fetcher.internal.filters import (
     AttributeFilter,
     Filter,
 )
